+++ CHANGELOG.md
@@ -65,6 +65,7 @@
 - `tracegen`: Add additional resource attributes (#11145)
 - `transformprocessor`: Add IsMatch factory function.  This function allows regex matching in conditions. (#10903)
 - `transformprocessor`: replace_pattern` and `replace_all_patterns` use regex for pattern matching and replacing text in attributes/metrics (#11125)
+- `internal/scrapertest` Add sortation functions for scrapertest (#10837)
 
 ### 🧰 Bug fixes 🧰
 - `aerospikereceiver`: Fix issue where namespaces would not be collected (#11465)